refactor(mcp): split proxy_http.go into per-transport sibling files
Pure file-moves; no behavior changes; no signature changes. Existing tests
on the mcp package stay green and coverage stays at 90.2%.

Files:
  proxy_http.go         (1901 -> 208 lines: header DLP helpers, RPC helpers, hashSessionKey, common types)
  mcp_http_forward.go   (RunHTTPProxy: stdio-to-HTTP forward bridge)
  mcp_http_reverse.go   (RunHTTPListenerProxy: HTTP reverse proxy server)
  mcp_http_input.go     (httpInputDecision, scanHTTPInput, scanHTTPInputDecision)
  mcp_sse_bridge.go     (sseMessageWriter, isSSEContentType, startGETStream)

5-file split (one more than the kickoff's 4-file prescription) because the
~770-line input-scanning chunk warrants its own file: it is shared by both
forward and reverse paths, has a single coherent purpose, and pulling it
out keeps proxy_http.go itself under the 500-line target instead of 977.

Each sibling file carries only the imports it needs after goimports prune.
Transport parity preserved: the same shared helpers serve forward, reverse,
and SSE bridge code paths exactly as before.
